Fix test
